Moving the test code to a proper file
